Assign list() instead of NULL to get empty list.
If we use `NULL' to create an empty list and then start assigning stuff
to single elements using `[[`, the order of assignments can make a
difference.  For example, if we first assign, say, a character vector of
length 1, the list will have `typeof' "character".  Therefore,
subsequent assignments of, say, a character vector of length >=2 will
fail.  If, on the other hand, our first assignment to the empty list is
a vector of length >=2, the list will automatically have `typeof'
"list".  Avoid this problem by making the intention of creating an empty
_list_ explicit via a call to `list()'.