Write extracted chunks directly to `outdir'.
Make the `idfile' mandatory.  Always write the order of individuals in
the extracted chunk files to a file called `order_of_individuals.txt' in
`outdir'.  Abort if such a file already exists.

Abort if we are about to create extracted chunk files that already
exists in `outdir'.

Let the perl script extract subsets of individuals.